requirements.txt
absl-py>=1.0.0
gin-config>=0.5.0
tensorflow>=2.1
tensorflow-datasets>=4.0.0
